+++ TestForSure/src/main/java/edu/tests/TestForSure/entity/TopPerformers.java
@@ -1,20 +1,22 @@
 package edu.tests.TestForSure.entity;
 
+import java.math.BigDecimal;
+
 public class TopPerformers {
 	private String name;
-	private int marks_scored;
+	private BigDecimal marks_scored;
 	private int rank;
-	private int time_taken;
+	private BigDecimal time_taken;
 	public String getName() {
 		return name;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getMarks_scored() {
+	public BigDecimal getMarks_scored() {
 		return marks_scored;
 	}
-	public void setMarks_scored(int marks_scored) {
+	public void setMarks_scored(BigDecimal marks_scored) {
 		this.marks_scored = marks_scored;
 	}
 	public int getRank() {
@@ -23,10 +25,10 @@ public class TopPerformers {
 	public void setRank(int rank) {
 		this.rank = rank;
 	}
-	public int getTime_taken() {
+	public BigDecimal getTime_taken() {
 		return time_taken;
 	}
-	public void setTime_taken(int time_taken) {
+	public void setTime_taken(BigDecimal time_taken) {
 		this.time_taken = time_taken;
 	}
 	@Override
@@ -34,7 +36,7 @@ public class TopPerformers {
 		return "TopPerformers [name=" + name + ", marks_scored=" + marks_scored + ", rank=" + rank + ", time_taken="
 				+ time_taken + "]";
 	}
-	public TopPerformers(String name, int marks_scored, int rank, int time_taken) {
+	public TopPerformers(String name, BigDecimal marks_scored, int rank, BigDecimal time_taken) {
 		super();
 		this.name = name;
 		this.marks_scored = marks_scored;
